rc-tracker-backend/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
djangorestframework==3.15
django-cors-headers==3.7.0  # Si tu as besoin de gérer les CORS pour la communication entre le frontend et le backend
django-environ==0.8.1  # Pour gérer les variables d'environnement (utile pour la configuration)
djangorestframework-simplejwt==5.2.2  # Pour la gestion de l'authentification JWT (si tu prévois d'utiliser JWT)
